+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 0.5.3dev
+* [Fix] Adds telemetry collect for `%sqlplot` invocation ([#89](https://github.com/ploomber/jupysql/pull/89))
 * [Fix] `setup.py` fix due to change in setuptools 67.0.0
 
 ## 0.5.2 (2023-01-03)